yes, i know i'm loosing precision, that's the point.

+++ common/logisticspipes/request/RequestTree.java
@@ -148,9 +148,9 @@ public class RequestTree extends RequestTreeNode {
 			}
 
 			//GetLoadFactor*64 should be an integer anyway.
-			c = Math.floor(o1.destination.getCachedPipe().getLoadFactor()*64) - Math.floor(o2.destination.getCachedPipe().getLoadFactor()*64);
+			c = (int)Math.floor(o1.destination.getCachedPipe().getLoadFactor()*64) - (int)Math.floor(o2.destination.getCachedPipe().getLoadFactor()*64);
 			if(distanceWeight != 0) {
-				c += (Math.floor(o1.distanceToDestination*64) - Math.floor(o2.distanceToDestination*64)) * distanceWeight;
+				c += (int)(Math.floor(o1.distanceToDestination*64) - (int)Math.floor(o2.distanceToDestination*64)) * distanceWeight;
 			}
 			/*
 			if (c < 0)
